Enum sage::asm::location::Location

pub enum Location {
    Address(usize),
    Indirect(Box<Self>),
    Offset(Box<Self>, isize),
}

A location in memory (on the tape of the virtual machine).

Variants

Address(usize)

A fixed position in the tape (a constant address known at compile time).

Indirect(Box<Self>)

Use the value of a cell on the tape as an address. For example, Indirect(SP) is the location of the top item on the stack.

Offset(Box<Self>, isize)

Go to a position in memory, and then move the pointer according to an offset. For example, Offset(Address(8), -2) is equivalent to Address(6).

Implementations

impl Location

pub fn offset(&self, offset: isize) -> Self

Get the location offset by a constant number of cells from a starting location. For example, Offset(Address(8), -2) is equivalent to Address(6).

pub fn deref(&self) -> Self

Get the location of the value pointed to by this location.

pub fn push_to(&self, sp: &Location, result: &mut dyn VirtualMachineProgram)

Push the value of this location to a given stack.

pub fn pop_from(&self, sp: &Location, result: &mut dyn VirtualMachineProgram)

Pop the top item off a given stack and store it in this location.

pub fn copy_address_to(&self, dst: &Self, result: &mut dyn VirtualMachineProgram)

Copy the address of this location to another location.

pub fn to(&self, result: &mut dyn VirtualMachineProgram)

Move the pointer to this location.

pub fn from(&self, result: &mut dyn VirtualMachineProgram)

Move the pointer from this location.

pub fn next(&self, count: isize, result: &mut dyn VirtualMachineProgram)

Take the pointer value of this location, and make it point count number of cells to the right of its original position.

pub fn prev(&self, count: isize, result: &mut dyn VirtualMachineProgram)

Take the pointer value of this location, and make it point count number of cells to the left of its original position.

pub fn whole_int(&self, result: &mut dyn VirtualMachineProgram)

Take the value at this location. If it is a whole number (>= 0), then the value of this location is now 1. Otherwise, the value is 0.

pub fn save_to(&self, result: &mut dyn VirtualMachineProgram)

Save the value of the virtual machine’s register to this location.

pub fn restore_from(&self, result: &mut dyn VirtualMachineProgram)

Restore the value from this location into the virtual machine’s register.

pub fn inc(&self, result: &mut dyn VirtualMachineProgram)

Increment the value of this location.

pub fn dec(&self, result: &mut dyn VirtualMachineProgram)

Decrement the value of this location.

pub fn bitwise_nand(&self, src: &Self, result: &mut dyn VirtualMachineProgram)

Perform bitwise-nand on this cell and a source cell.

pub fn not(&self, result: &mut dyn VirtualMachineProgram)

If this cell is non-zero, then the value of this location is now 0. Otherwise, the value of this location is now 1.

Perform boolean not on the value of this cell

pub fn and(&self, src: &Self, result: &mut dyn VirtualMachineProgram)

Perform boolean and on the value of this cell and a source cell.

pub fn or(&self, src: &Self, result: &mut dyn VirtualMachineProgram)

Perform boolean or on the value of this cell and a source cell.

pub fn is_greater_than(
&self,
src: &Self,
dst: &Self,
result: &mut dyn VirtualMachineProgram
)

dst = this cell > source cell.

pub fn is_greater_than_float(
&self,
src: &Self,
dst: &Self,
result: &mut dyn VirtualMachineProgram
) -> Result<(), Error>

dst = this cell >= source cell.

pub fn is_greater_or_equal_to(
&self,
src: &Self,
dst: &Self,
result: &mut dyn VirtualMachineProgram
)

dst = this cell >= source cell.

pub fn is_less_than(
&self,
src: &Self,
dst: &Self,
result: &mut dyn VirtualMachineProgram
)

dst = this cell < source cell.

pub fn is_less_than_float(
&self,
src: &Self,
dst: &Self,
result: &mut dyn VirtualMachineProgram
) -> Result<(), Error>

dst = this cell < source cell.

pub fn is_less_or_equal_to(
&self,
src: &Self,
dst: &Self,
result: &mut dyn VirtualMachineProgram
)

dst = this cell <= source cell.

pub fn is_not_equal(
&self,
src: &Self,
dst: &Self,
result: &mut dyn VirtualMachineProgram
)

pub fn is_equal(
&self,
src: &Self,
dst: &Self,
result: &mut dyn VirtualMachineProgram
)

pub fn add(&self, src: &Self, result: &mut dyn VirtualMachineProgram)

This cell += source cell.

pub fn sub(&self, src: &Self, result: &mut dyn VirtualMachineProgram)

This cell -= source cell.

pub fn mul(&self, src: &Self, result: &mut dyn VirtualMachineProgram)

This cell *= source cell.

pub fn div(&self, src: &Self, result: &mut dyn VirtualMachineProgram)

This cell /= source cell.

pub fn rem(&self, src: &Self, result: &mut dyn VirtualMachineProgram)

This cell %= source cell.

pub fn add_float(
&self,
src: &Self,
result: &mut dyn VirtualMachineProgram
) -> Result<(), Error>

This cell += source cell.

pub fn sub_float(
&self,
src: &Self,
result: &mut dyn VirtualMachineProgram
) -> Result<(), Error>

This cell -= source cell.

pub fn mul_float(
&self,
src: &Self,
result: &mut dyn VirtualMachineProgram
) -> Result<(), Error>

This cell *= source cell.

pub fn div_float(
&self,
src: &Self,
result: &mut dyn VirtualMachineProgram
) -> Result<(), Error>

This cell /= source cell.

pub fn rem_float(
&self,
src: &Self,
result: &mut dyn VirtualMachineProgram
) -> Result<(), Error>

This cell %= source cell.

pub fn pow_float(
&self,
src: &Self,
result: &mut dyn VirtualMachineProgram
) -> Result<(), Error>

This cell **= source cell.

pub fn set(&self, val: isize, result: &mut dyn VirtualMachineProgram)

This cell = a constant value.

pub fn set_float(
&self,
val: f64,
result: &mut dyn VirtualMachineProgram
) -> Result<(), Error>

This cell = a constant floating point value. This requires the standard instruction.

pub fn alloc(&self, result: &mut dyn VirtualMachineProgram) -> Result<(), Error>

Read the value of this cell, allocate that number of cells, store the address of the first cell in this cell.

pub fn free(&self, result: &mut dyn VirtualMachineProgram) -> Result<(), Error>

Free the pointer stored in this cell, and set the value to -1000 (to prevent) accidental use of the freed memory.

pub fn sin(&self, result: &mut dyn VirtualMachineProgram) -> Result<(), Error>

pub fn cos(&self, result: &mut dyn VirtualMachineProgram) -> Result<(), Error>

pub fn tan(&self, result: &mut dyn VirtualMachineProgram) -> Result<(), Error>

pub fn asin(&self, result: &mut dyn VirtualMachineProgram) -> Result<(), Error>

pub fn acos(&self, result: &mut dyn VirtualMachineProgram) -> Result<(), Error>

pub fn atan(&self, result: &mut dyn VirtualMachineProgram) -> Result<(), Error>

pub fn to_float(
&self,
result: &mut dyn VirtualMachineProgram
) -> Result<(), Error>

pub fn to_int(&self, result: &mut dyn VirtualMachineProgram) -> Result<(), Error>

pub fn get(&self, src: Input, result: &mut dyn VirtualMachineProgram)

pub fn put(&self, dst: Output, result: &mut dyn VirtualMachineProgram)

pub fn peek(&self, result: &mut dyn VirtualMachineProgram) -> Result<(), Error>

pub fn poke(&self, result: &mut dyn VirtualMachineProgram) -> Result<(), Error>

pub fn copy_to(&self, dst: &Self, result: &mut dyn VirtualMachineProgram)

Store the value of this cell into another cell.