sage::asm
Module std

Standard Assembly Variant

This variant of the assembly language is intended to be used with the standard variant of the virtual machine. It is very portable, but probably not supported on older systems or hardware implementations.

Click here to view opcodes!

Structs

StandardProgram

A program composed of standard instructions, which can be assembled into the standard virtual machine instructions.

Enums

StandardOp

A standard instruction of the assembly language. These are instructions that should be implemented for every target possible. Standard instructions should only not be implemented for targets like physical hardware, where the program is executed on the bare metal (a custom CPU or FPGA).