Enum sage::asm::std::StandardOp

pub enum StandardOp {

    CoreOp(CoreOp),
    Set(Location, f64),
    ToFloat(Location),
    ToInt(Location),
    Pow {
        src: Location,
        dst: Location,
    },
    Sqrt(Location),
    Add {
        src: Location,
        dst: Location,
    },
    Sub {
        src: Location,
        dst: Location,
    },
    Mul {
        src: Location,
        dst: Location,
    },
    Div {
        src: Location,
        dst: Location,
    },
    Rem {
        src: Location,
        dst: Location,
    },
    Neg(Location),
    Sin(Location),
    Cos(Location),
    Tan(Location),
    ASin(Location),
    ACos(Location),
    ATan(Location),
    IsGreater {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsLess {
        a: Location,
        b: Location,
        dst: Location,
    },
    Alloc(Location),
    Free(Location),
    Peek(Location),
    Poke(Location),

}

A standard instruction of the assembly language. These are instructions that should be implemented for every target possible. Standard instructions should only not be implemented for targets like physical hardware, where the program is executed on the bare metal (a custom CPU or FPGA).

Variants

CoreOp(CoreOp)

Execute a core instruction.

Set(Location, f64)

Set the value of a cell to a constant float.

ToFloat(Location)

Take the integer value stored in a cell and store the equivalent float value in the same cell.

ToInt(Location)

Take the float value stored in a cell and store the equivalent integer value in the same cell.

Pow

Fields

src: Location

The source cell.

dst: Location

The destination cell.

Raise a cell (float) to the power of another cell (float).

Sqrt(Location)

Take the square root of a cell (float).

Add

Fields

src: Location

The source cell.

dst: Location

The destination cell.

Add the source cell (float) to the destination cell (float).

Sub

Fields

src: Location

The source cell.

dst: Location

The destination cell.

Subtract the source cell (float) from the destination cell (float).

Mul

Fields

src: Location

The source cell.

dst: Location

The destination cell.

Multiply the source cell (float) by the destination cell (float).

Div

Fields

src: Location

The source cell.

dst: Location

The destination cell.

Divide the destination cell (float) by the source cell (float).

Rem

Fields

src: Location

The source cell.

dst: Location

The destination cell.

Perform the modulo operation on the destination cell (float) by the source cell (float).

Neg(Location)

Negate the value of a cell (float) and store the result in the same cell.

Sin(Location)

Perform Sin on a cell (float) and store the result in the same cell.

Cos(Location)

Perform Cos on a cell (float) and store the result in the same cell.

Tan(Location)

Perform Tan on a cell (float) and store the result in the same cell.

ASin(Location)

Perform inverse Sin on a cell (float) and store the result in the same cell.

ACos(Location)

Perform inverse Cos on a cell (float) and store the result in the same cell.

ATan(Location)

Perform inverse Tan on a cell (float) and store the result in the same cell.

IsGreater

Fields

a: Location

The first cell in the comparison (left hand side).

b: Location

The second cell in the comparison (right hand side).

dst: Location

The destination cell.

Perform dst = a > b.

IsLess

Fields

a: Location

The first cell in the comparison (left hand side).

b: Location

The second cell in the comparison (right hand side).

dst: Location

The destination cell.

Perform dst = a < b.

Alloc(Location)

Take the value in the operand cell. Allocate that number of cells. The address of the first cell is stored in the operand cell.

Free(Location)

Free the memory allocated at the address stored in the operand cell.

Peek(Location)

Read from the runtime’s interface into the operand cell.

Poke(Location)

Write from the operand cell into the runtime’s interface.