sage::asm
Trait AssemblyProgram

pub trait AssemblyProgram {
    // Required methods
    fn op(&mut self, op: CoreOp);
    fn std_op(&mut self, op: StandardOp) -> Result<(), Error>;
    fn is_defined(&self, label: &str) -> bool;
    fn current_instruction(&self) -> usize;
    fn get_op(&self, start: usize) -> Option<Result<CoreOp, StandardOp>>;

    // Provided methods
    fn comment(&mut self, comment: String) { ... }
    fn log_instructions_after(&self, target: &str, message: &str, i: usize) { ... }
}

A frontend to both the CoreProgram and StandardProgram types. This allows the compiler to append CoreOps to both programs with guaranteed success, but optionally allows the compiler to attempt to compile a StandardOp, which may fail depending on the target’s support for the instruction.

Required Methods

fn op(&mut self, op: CoreOp)

Insert a core operation into the program.

fn std_op(&mut self, op: StandardOp) -> Result<(), Error>

Attempt to insert a standard operation into the program. This could fail depending on the backend’s support for the instruction.

fn is_defined(&self, label: &str) -> bool

Is the given label defined yet in the operations? I.E., has a CoreOp::Fn with this label been inserted into the program code yet?

fn current_instruction(&self) -> usize

Get the current instruction number.

fn get_op(&self, start: usize) -> Option<Result<CoreOp, StandardOp>>

Get the operation at the given instruction number.

Provided Methods

fn comment(&mut self, comment: String)

Insert a comment into the program.

fn log_instructions_after(&self, target: &str, message: &str, i: usize)

Log all the instructions after the given instruction number. The target is an identifier for the instructions being logged. This is used to search for the log messages in the output. The message is a message to be logged before the instructions. This tells some context about the instructions being logged. The i is the instruction number to start logging at. This will log all instructions until the end of the program.