Enum sage::lir::ConstExpr

pub enum ConstExpr {

    Annotated(Box<Self>, Annotation),
    Declare(Box<Declaration>, Box<Self>),
    None,
    Null,
    Symbol(String),
    Int(i64),
    Cell(i64),
    Float(f64),
    Char(char),
    Bool(bool),
    Of(Type, String),
    TypeOf(Box<Expr>),
    SizeOfType(Type),
    SizeOfExpr(Box<Expr>),
    Type(Type),
    Tuple(Vec<Self>),
    Array(Vec<Self>),
    Struct(BTreeMap<String, Self>),
    Union(Type, String, Box<Self>),
    EnumUnion(Type, String, Box<Self>),
    CoreBuiltin(CoreBuiltin),
    StandardBuiltin(StandardBuiltin),
    FFIProcedure(FFIProcedure),
    Proc(Procedure),
    PolyProc(PolyProcedure),
    Monomorphize(Box<Self>, Vec<Type>),
    Template(Vec<String>, Box<Self>),
    Member(Box<Self>, Box<Self>),
    As(Box<Self>, Type),

}

A compiletime expression.

Variants

Annotated(Box<Self>, Annotation)

Declare(Box<Declaration>, Box<Self>)

Bind a list of types in a constant expression.

None

The unit, or “void” instance.

Null

The null pointer constant.

Symbol(String)

A named constant.

Int(i64)

A constant integer value.

Cell(i64)

A constant integer value representing a cell on the tape.

Float(f64)

A constant floating point value.

These can be used at compile time even when compiling to core, but are not allowed at runtime in the core variant.

Char(char)

A constant chararacter.

Bool(bool)

A constant boolean value.

Of(Type, String)

A constant enum variant.

TypeOf(Box<Expr>)

Get the type of an expression. (as an array of chars)

SizeOfType(Type)

Get the size of a type (in cells) as a constant int.

SizeOfExpr(Box<Expr>)

Get the size of an expression’s type (in cells) as a constant int. This will not evaluate the inner expression.

Type(Type)

A type as a constant expression.

Tuple(Vec<Self>)

A tuple of constant values.

Array(Vec<Self>)

An array of constant values.

Struct(BTreeMap<String, Self>)

A structure of constant values.

Union(Type, String, Box<Self>)

A union of constant values.

EnumUnion(Type, String, Box<Self>)

A tagged union of constant values.

CoreBuiltin(CoreBuiltin)

A builtin implemented in handwritten core assembly.

StandardBuiltin(StandardBuiltin)

A builtin implemented in handwritten standard assembly.

FFIProcedure(FFIProcedure)

A foreign function interface binding.

Proc(Procedure)

A procedure.

PolyProc(PolyProcedure)

A polymorphic procedure.

Monomorphize(Box<Self>, Vec<Type>)

Monomorphize a constant expression with some type arguments.

Template(Vec<String>, Box<Self>)

Member(Box<Self>, Box<Self>)

Get an attribute of a constant expression.

As(Box<Self>, Type)

Cast a constant expression to another type.

Implementations

impl ConstExpr

pub fn monomorphize(self, ty_args: Vec<Type>) -> Self

pub fn template(&self, params: Vec<String>) -> Self

pub fn proc( common_name: Option<String>, args: Vec<(String, Mutability, Type)>, ret: Type, body: impl Into<Expr> ) -> Self

Construct a procedure.

pub fn with(&self, older_decls: impl Into<Declaration>) -> Self

Return this expression, but with a given declaration in scope.

pub fn annotate(self, annotation: Annotation) -> Self

Annotate this constant expression with a source code location.

pub fn app(self, args: Vec<Expr>) -> Expr

Apply this procedure or builtin to a list of expressions at runtime.

pub fn eval(self, env: &Env) -> Result<Self, Error>

Evaluate this constant expression at compile time, and get the result.

pub fn as_type(self, t: Type) -> Self

Cast an expression as another type.

pub fn field(self, field: ConstExpr) -> Self

Get a field from a structure, union, or tuple.

For tuples, use an Int constant expression to access the nth field (zero indexed). For unions or structures, use a Symbol constant expression to access the field.

pub fn as_int(self, env: &Env) -> Result<i64, Error>

Try to get this constant expression as an integer.

pub fn as_float(self, env: &Env) -> Result<f64, Error>

Try to get this constant expression as a float.

pub fn as_bool(self, env: &Env) -> Result<bool, Error>

Try to get this constant expression as a boolean value.

pub fn as_symbol(self, env: &Env) -> Result<String, Error>

Try to get this constant expression as a symbol (like in LISP).