Enum sage::lir::Error

pub enum Error {

    AssemblyError(Error),
    VariantNotFound(Type, String),
    MemberNotFound(Expr, ConstExpr),
    RecursionDepthConst(ConstExpr),
    CouldntSimplify(Type, Type),
    RecursionDepthTypeEquality(Type, Type),
    NonIntegralConst(ConstExpr),
    UnsizedType(Type),
    DerefNonPointer(Expr),
    ApplyNonProc(Expr),
    NonSymbol(ConstExpr),
    InvalidIndex(Expr),
    InvalidRefer(Expr),
    InvalidUnaryOp(Box<dyn UnaryOp>, Expr),
    InvalidUnaryOpTypes(Box<dyn UnaryOp>, Type),
    InvalidBinaryOp(Box<dyn BinaryOp>, Expr, Expr),
    InvalidBinaryOpTypes(Box<dyn BinaryOp>, Type, Type),
    InvalidTernaryOp(Box<dyn TernaryOp>, Expr, Expr, Expr),
    InvalidTernaryOpTypes(Box<dyn TernaryOp>, Type, Type, Type),
    InvalidAssignOp(Box<dyn AssignOp>, Expr, Expr),
    InvalidAssignOpTypes(Box<dyn AssignOp>, Type, Type),
    MismatchedTypes {
        expected: Type,
        found: Type,
        expr: Expr,
    },
    SymbolNotDefined(String),
    TypeNotDefined(String),
    NegativeArrayLength(Expr),
    InvalidPatternForType(Type, Pattern),
    InvalidPatternForExpr(Expr, Pattern),
    InvalidMatchExpr(Expr),
    NonExhaustivePatterns {
        patterns: Vec<Pattern>,
        expr: Expr,
    },
    InvalidAs(Expr, Type, Type),
    InvalidConstExpr(ConstExpr),
    UnsupportedOperation(Expr),
    TypeRedefined(String),
    UnusedExpr(Expr, Type),
    InvalidTemplateArgs(Type),
    ApplyNonTemplate(Type),
    SizeOfTemplate(Type),
    CompilePolyProc(PolyProcedure),
    InvalidMonomorphize(ConstExpr),

}

An LIR compilation error.

Variants

AssemblyError(Error)

An error caused by trying to assemble invalid code generated by the compiler. This should be taken seriously, unless the error is due to an invalid handwritten builtin.

VariantNotFound(Type, String)

The variant of an enum is not defined.

MemberNotFound(Expr, ConstExpr)

Tried to access an undefined member of a tuple, struct, or union.

RecursionDepthConst(ConstExpr)

Recursion depth exceeded when trying to evaluate a constant expression.

CouldntSimplify(Type, Type)

Recursion depth exceeded when trying to confirm a type’s equality to another type.

RecursionDepthTypeEquality(Type, Type)

NonIntegralConst(ConstExpr)

Got another type when expecting an integer, bool, or char.

UnsizedType(Type)

Tried to instantiate a type that cannot be sized. This is a problem because we cannot manage the stack if we cannot know the size of the type.

DerefNonPointer(Expr)

Tried to dereference a non-pointer.

ApplyNonProc(Expr)

Tried to apply a non-procedure to some arguments.

NonSymbol(ConstExpr)

Expected a symbol, but got something else.

InvalidIndex(Expr)

Invalid Index expression (incorrect types).

InvalidRefer(Expr)

Invalid Refer expression. The compiler was not able to calculate the address of the expression.

InvalidUnaryOp(Box<dyn UnaryOp>, Expr)

Invalid unary operation (negate, not) expression (incorrect types).

InvalidUnaryOpTypes(Box<dyn UnaryOp>, Type)

Invalid unary op types (incorrect types).

InvalidBinaryOp(Box<dyn BinaryOp>, Expr, Expr)

Invalid binary operation (add, subtract, and, or) expression (incorrect types).

InvalidBinaryOpTypes(Box<dyn BinaryOp>, Type, Type)

Invalid binary op types (incorrect types).

InvalidTernaryOp(Box<dyn TernaryOp>, Expr, Expr, Expr)

Invalid ternary operation (if) expression (incorrect types).

InvalidTernaryOpTypes(Box<dyn TernaryOp>, Type, Type, Type)

Invalid ternary op types (incorrect types).

InvalidAssignOp(Box<dyn AssignOp>, Expr, Expr)

Invalid assignment operation (assign, add_assign, subtract_assign, and_assign, or_assign) expression (incorrect types).

InvalidAssignOpTypes(Box<dyn AssignOp>, Type, Type)

Invalid assign op types (incorrect types).

MismatchedTypes

Fields

expected: Type

found: Type

expr: Expr

Mismatched types

SymbolNotDefined(String)

A symbol was used, but not defined.

TypeNotDefined(String)

A type was used, but not defined.

NegativeArrayLength(Expr)

Tried to create an array with a negative length.

InvalidPatternForType(Type, Pattern)

Tried to use a pattern that is not valid for the given type.

InvalidPatternForExpr(Expr, Pattern)

Tried to use a pattern that is not valid for the given expression.

InvalidMatchExpr(Expr)

Tried to match over an expression that cannot be matched over.

NonExhaustivePatterns

Fields

patterns: Vec<Pattern>

expr: Expr

Invalid pattern for a match expression.

InvalidAs(Expr, Type, Type)

Invalid type casting expression.

InvalidConstExpr(ConstExpr)

Invalid constant expression.

UnsupportedOperation(Expr)

Expression uses an operation unsupported by the target.

TypeRedefined(String)

Tried to define a type that already exists.

UnusedExpr(Expr, Type)

Unused expression returned a non-None value.

InvalidTemplateArgs(Type)

Invalid number of template arguments to a type.

ApplyNonTemplate(Type)

Tried to apply a non-template type to some arguments.

SizeOfTemplate(Type)

Tried to get the size of a template type.

CompilePolyProc(PolyProcedure)

Tried to compile a polymorphic procedure without monomorphing it.

InvalidMonomorphize(ConstExpr)

Cannot monomorphize a constant expression.