Enum sage::lir::Pattern

pub enum Pattern {
    Tuple(Vec<Pattern>),
    Struct(HashMap<String, Pattern>),
    Variant(String, Option<Box<Pattern>>),
    Symbol(Mutability, String),
    ConstExpr(ConstExpr),
    Alt(Vec<Pattern>),
    Pointer(Box<Pattern>),
    Wildcard,
}

A pattern which can be matched against an expression.

Patterns are used in the match expression, and in the let expression. Patterns can either match or bind to an expression. When matching, the pattern just checks if the expression matches the pattern. When binding, the pattern binds the expression to corresponding variables in the pattern, and evaluates an expression with those variables.

Variants

Tuple(Vec<Pattern>)

Struct(HashMap<String, Pattern>)

Variant(String, Option<Box<Pattern>>)

Symbol(Mutability, String)

ConstExpr(ConstExpr)

Alt(Vec<Pattern>)

Pointer(Box<Pattern>)

Wildcard

Implementations

impl Pattern

pub fn tup(patterns: Vec<Pattern>) -> Self

Construct a new pattern which matches a tuple of patterns.

pub fn struct_(patterns: HashMap<String, Pattern>) -> Self

Construct a new pattern which matches a struct with a given set of fields.

pub fn sym(mutable: impl Into<Mutability>, name: impl ToString) -> Self

Construct a new pattern which matches a symbol with a given name.

pub fn int(n: i64) -> Self

Construct a new pattern which matches a constant integer.

pub fn float(n: f64) -> Self

Construct a new pattern which matches a constant float.

pub fn bool(b: bool) -> Self

Construct a new pattern which matches a constant boolean.

pub fn alt(patterns: Vec<Pattern>) -> Self

Construct a new pattern which binds to several alternate patterns.

pub fn wildcard() -> Self

Construct a new pattern which matches any expression.

pub fn pointer(pattern: Pattern) -> Self

Construct a new pattern which matches a pointer.

pub fn get_branch_result_type( &self, expr: &Expr, branch: &Expr, env: &Env ) -> Result<Type, Error>

Get the type of a branch with a given expression matched to this pattern.

pub fn are_patterns_exhaustive( expr: &Expr, patterns: &[Pattern], matching_expr_ty: &Type, env: &Env ) -> Result<bool, Error>

This associated function returns whether or not a set of patterns is exhaustive, that is, whether or not it matches all possible values of a given type. This is used to check if a match expression is exhaustive.

pub fn type_check( &self, matching_expr: &Expr, branch: &Expr, env: &Env ) -> Result<(), Error>

Type-check a pattern match of an expression against this pattern, and type-check the branch where the expression is bound to the pattern.

pub fn if_let_pattern( &self, expr: &Expr, then: &Expr, else_: &Expr, env: &Env ) -> Result<Expr, Error>

Generate an if letexpression, which matches a given expr, and executes then if the expression matches the pattern, and else_ otherwise.

pub fn match_pattern( expr: &Expr, branches: &[(Self, Expr)], env: &Env ) -> Result<Expr, Error>

Generate an expression which evaluates a match expression, which matches a given expr against a set of patterns and branches.

pub fn get_bindings( &self, expr: &Expr, ty: &Type, env: &Env ) -> Result<HashMap<String, (Mutability, Type)>, Error>

Get the map of new variables and their types which are bound by this pattern.

For example (a, b) binds the variables a and b to the first and second elements of the tuple respectively. This function would return a map with the keys a and b and the values being their types.