Struct sage::lir::CoreBuiltin

pub struct CoreBuiltin {
    pub name: String,
    pub args: Vec<(String, Type)>,
    pub ret: Type,
    pub body: Vec<CoreOp>,
}

A builtin pseudo-procedure implemented in the core assembly variant.

This is not actually executed like a legitimate procedure, but instead the assembly code is directly pasted where the procedure is called.

The arguments of the procedure are pushed to the stack in the order they are given. The builtin must pop its arguments. The return value is left on the stack in their place by the builtin.

Fields
name: String

The name of the builtin. This isn’t used in compilation, just for debugging.

args: Vec<(String, Type)>

The arguments of the builtin. These will be typechecked when the builtin is called.

ret: Type

The return value the builtin will leave on the stack after popping its arguments.

body: Vec<CoreOp>

The list of assembly instructions to be pasted into the assembly code.