sage::lir
Struct Env

pub struct Env { /* private fields */ }

An environment under which expressions and types are compiled and typechecked. This is essentially the scope of an expression.

Implementations

impl Env

pub fn get_type_of_associated_const( &self, ty: &Type, name: &str, ) -> Option<Type>

Get the type of an associated constant of a type.

pub fn get_associated_const( &self, ty: &Type, name: &str, ) -> Option<(ConstExpr, Type)>

pub fn has_associated_const(&self, ty: &Type, name: &str) -> bool

pub fn get_all_associated_consts(&self, ty: &Type) -> Vec<(String, ConstExpr)>

pub fn add_associated_const( &self, ty: Type, associated_const_name: impl ToString, expr: ConstExpr, ) -> Result<(), Error>

pub fn define_types(&mut self, types: Vec<(String, Type)>)

Define multiple types with the given names under this environment.

This must be used in situations where the different types depend on each other. This is because the sizes of types are memoized, and this will interfere with the memoization process if the types are defined separately. It will lead to typechecking errors if the environment does not already have a memoized size for the type of a subexpression.

pub fn define_var( &mut self, var: impl ToString, mutability: Mutability, ty: Type, compiling: bool, ) -> Result<isize, Error>

Define a variable in the current scope. This will increment the scopeโs frame pointer offset by the size of the variable. This method returns the offset of the variable from the frame pointer under this scope.

pub fn compile_args( &mut self, args: impl IntoIterator<Item = Expr>, output: &mut dyn AssemblyProgram, ) -> Result<(), Error>

Compile a list of arguments, accounting for the fact that the stack is growing as we compile arguments.

Use this to compile arguments to an operator, to a function, or subexpressions of an intermediate expression.

Statement expressions donโt need to use this method because they donโt need to save the intermediate states of the stack.