Struct sage::lir::Procedure

pub struct Procedure { /* private fields */ }

A monomorphic procedure of LIR code which can be applied to a list of arguments. A procedure is compiled down to a label in the assembly code. The label is called with the Call instruction.

Implementations

impl Procedure

pub fn new( common_name: Option<String>, args: Vec<(String, Mutability, Type)>, ret: Type, body: impl Into<Expr> ) -> Self

Construct a new procedure with a given list of arguments and their types, a return type, and the body of the procedure.

pub fn with(&self, decls: impl Into<Declaration>) -> Self

pub fn get_args(&self) -> &[(String, Mutability, Type)]

Get the arguments of the procedure.

pub fn get_ret(&self) -> &Type

Get the return type of the procedure.

pub fn get_body(&self) -> &Expr

Get the body of the procedure.

pub fn get_mangled_name(&self) -> &str

Get the mangled name of the procedure. The procedure’s mangled name is used to store the procedure in the environment.

pub fn get_common_name(&self) -> Option<&str>

Get the name of the procedure known to the LIR front-end.

pub fn set_common_name(&mut self, name: impl ToString)

This is just for debugging purposes. This sets the common name of the procedure known to the user, or the front-end of the compiler.

pub fn push_label(&self, output: &mut dyn AssemblyProgram)

Push this procedure’s label to the stack.