Trait sage::lir::Simplify

pub trait Simplify: Sized {
    fn simplify_checked(self, env: &Env, i: usize) -> Result<Self, Error>;

    fn simplify(self, env: &Env) -> Result<Self, Error> { ... }
}

Simplify an expression while maintaining structural equality.

Required Methods

fn simplify_checked(self, env: &Env, i: usize) -> Result<Self, Error>

Simplify an expression while maintaining structural equality, but limit the number of recursive calls to prevent stack overflow. i is a counter to prevent infinite recursion.

Provided Methods

fn simplify(self, env: &Env) -> Result<Self, Error>

Simplify an expression while maintaining structural equality.