Trait sage::lir::TernaryOp

pub trait TernaryOp: Debug + Display {
    // Required methods
    fn clone_box(&self) -> Box<dyn TernaryOp>;
    fn can_apply(
        &self,
        a: &Type,
        b: &Type,
        c: &Type,
        env: &Env
    ) -> Result<bool, Error>;
    fn eval(
        &self,
        a: &ConstExpr,
        b: &ConstExpr,
        c: &ConstExpr,
        env: &mut Env
    ) -> Result<ConstExpr, Error>;
    fn compile_types(
        &self,
        a: &Type,
        b: &Type,
        c: &Type,
        env: &mut Env,
        output: &mut dyn AssemblyProgram
    ) -> Result<(), Error>;

    // Provided methods
    fn type_check(
        &self,
        a: &Expr,
        b: &Expr,
        c: &Expr,
        env: &Env
    ) -> Result<(), Error> { ... }
    fn return_type(
        &self,
        a: &Expr,
        b: &Expr,
        c: &Expr,
        env: &Env
    ) -> Result<Type, Error> { ... }
    fn display(&self, a: &Expr, b: &Expr, c: &Expr) -> String { ... }
    fn can_apply_exprs(
        &self,
        a: &Expr,
        b: &Expr,
        c: &Expr,
        env: &Env
    ) -> Result<bool, Error> { ... }
    fn compile(
        &self,
        a: &Expr,
        b: &Expr,
        c: &Expr,
        env: &mut Env,
        output: &mut dyn AssemblyProgram
    ) -> Result<(), Error> { ... }
}

A trait used to implement a ternary operation.

Required Methods

fn clone_box(&self) -> Box<dyn TernaryOp>

Clones the operation into a boxed trait object.

fn can_apply( &self, a: &Type, b: &Type, c: &Type, env: &Env ) -> Result<bool, Error>

Checks if the operation can be applied to the given types.

fn eval( &self, a: &ConstExpr, b: &ConstExpr, c: &ConstExpr, env: &mut Env ) -> Result<ConstExpr, Error>

Evaluates the operation on the given constant expressions.

fn compile_types( &self, a: &Type, b: &Type, c: &Type, env: &mut Env, output: &mut dyn AssemblyProgram ) -> Result<(), Error>

Compiles the operation on the given types. (Generates the code for the operation.)

Provided Methods

fn type_check( &self, a: &Expr, b: &Expr, c: &Expr, env: &Env ) -> Result<(), Error>

Typechecks the operation on the given expressions.

fn return_type( &self, a: &Expr, b: &Expr, c: &Expr, env: &Env ) -> Result<Type, Error>

Gets the type of the operation on the given expressions.

fn display(&self, a: &Expr, b: &Expr, c: &Expr) -> String

Formats the operation for display.

fn can_apply_exprs( &self, a: &Expr, b: &Expr, c: &Expr, env: &Env ) -> Result<bool, Error>

Checks if the operation can be applied to the given expressions.

fn compile( &self, a: &Expr, b: &Expr, c: &Expr, env: &mut Env, output: &mut dyn AssemblyProgram ) -> Result<(), Error>

Compiles the operation on the given expressions.