Trait sage::lir::UnaryOp

pub trait UnaryOp: Debug + Display {
    // Required methods
    fn clone_box(&self) -> Box<dyn UnaryOp>;
    fn can_apply(&self, expr: &Type, env: &Env) -> Result<bool, Error>;
    fn eval(&self, expr: &ConstExpr, env: &mut Env) -> Result<ConstExpr, Error>;
    fn compile_types(
        &self,
        expr: &Type,
        env: &mut Env,
        output: &mut dyn AssemblyProgram
    ) -> Result<(), Error>;

    // Provided methods
    fn type_check(&self, expr: &Expr, env: &Env) -> Result<(), Error> { ... }
    fn return_type(&self, expr: &Expr, env: &Env) -> Result<Type, Error> { ... }
    fn display(&self, expr: &Expr) -> String { ... }
    fn can_apply_exprs(&self, expr: &Expr, env: &Env) -> Result<bool, Error> { ... }
    fn compile(
        &self,
        expr: &Expr,
        env: &mut Env,
        output: &mut dyn AssemblyProgram
    ) -> Result<(), Error> { ... }
}

A trait used to implement a unary operation.

This trait is used to implement unary operations like not and ~.

Required Methods

fn clone_box(&self) -> Box<dyn UnaryOp>

Clones the operation into a boxed trait object.

fn can_apply(&self, expr: &Type, env: &Env) -> Result<bool, Error>

Checks if the operation can be applied to the given type.

fn eval(&self, expr: &ConstExpr, env: &mut Env) -> Result<ConstExpr, Error>

Evaluates the operation on the given constant expression.

fn compile_types( &self, expr: &Type, env: &mut Env, output: &mut dyn AssemblyProgram ) -> Result<(), Error>

Compiles the operation on the given type. (Generates the code for the operation.)

Provided Methods

fn type_check(&self, expr: &Expr, env: &Env) -> Result<(), Error>

Typechecks the operation on the given expression.

fn return_type(&self, expr: &Expr, env: &Env) -> Result<Type, Error>

Gets the type of the operation on the given expression.

fn display(&self, expr: &Expr) -> String

Formats the operation for display.

fn can_apply_exprs(&self, expr: &Expr, env: &Env) -> Result<bool, Error>

Checks if the operation can be applied to the given expression.

fn compile( &self, expr: &Expr, env: &mut Env, output: &mut dyn AssemblyProgram ) -> Result<(), Error>

Compiles the operation on the given expression.