sage
Module parse

Parsing Module

This module contains all the related functions for parsing any given source language supported by the compiler.

Right now, this module supports parsing:

LIR source

Core Assembly source

Standard Assembly source

Core Virtual Machine code source

Standard Virtual Machine code source

Stability

This module will likely be unstable for a long while, because none of the language syntaxes are stablized.

I also hope to replace the parsers with libraries using fewer dependencies. LALRPOP requires a ridiculous number of packages to function. Nom seems to be an ideal candidate.

Error Types

The error types in this module are just strings for now, but I intend to replace these with a set of full blown syntax error enums in the future.

Structs

SourceCodeLocation

A struct representing a location in the source code. This is used to format errors properly.

Functions

parse_asm

Parse Core and Standard variants of assembly source code. This will return core code by default, but will fallback on standard.

parse_frontend

Parse frontend sage code into an LIR expression.

parse_frontend_minimal

parse_lir

Parse LIR code as an LIR expression.

parse_vm

Parse Core and Standard variants of virtual machine source code. This will return core code by default, but will fallback on standard.