sage::side_effects
Module io

I/O Module

This module implements all the types related to the I/O operations available to the core instruction set. This allows the core VM to communicate with various instruments connected to the VM, without directly calling foreign functions. This is a standardized interface for various useful types of I/O, that each platform can implement in their own way, if they so choose.

Structs

Channel

The channel to use for a given I/O mode.

Input

An input source for a program.

Output

An output destination for a program.

Enums

Axis

The different axes an input or output might use.

Color

The different output colors a program might use.

Direction

The different directions a D-Pad a might use.

InputMode

The different types of input modes a program might use.

OutputMode

The different types of output modes a program might use.