sage::targets::sage_lisp
Struct SageLisp

pub struct SageLisp(/* private fields */);

The type for the C target which implements the Target trait. This allows the compiler to target the C language.

Implementations

impl SageLisp

pub fn new(env: Env) -> Self