sage::targets
Trait Architecture

pub trait Architecture {

    // Required methods
    fn name(&self) -> &str;
    fn version(&self) -> &str;
    fn supports_floats(&self) -> bool;
    fn supports_input(&self, src: &Input) -> bool;
    fn supports_output(&self, dst: &Output) -> bool;
    fn get(&mut self, src: &Input) -> Result<String, String>;
    fn put(&mut self, dst: &Output) -> Result<String, String>;
    fn peek(&mut self) -> Result<String, String>;
    fn poke(&mut self) -> Result<String, String>;
    fn declare_proc(&mut self, label_id: usize) -> String;
    fn end(&mut self, matching: &CoreOp, fun: Option<usize>) -> String;
    fn op(&mut self, op: &CoreOp) -> String;
    fn std_op(&mut self, op: &StandardOp) -> Result<String, String>;

    // Provided methods
    fn prelude(&self, _is_core: bool) -> Option<String> { ... }
    fn postop(&self) -> Option<String> { ... }
    fn postlude(&self, _is_core: bool) -> Option<String> { ... }
    fn pre_funs(&self, _funs: Vec<i32>) -> Option<String> { ... }
    fn post_funs(&self, _funs: Vec<i32>) -> Option<String> { ... }
    fn indentation(&self) -> Option<String> { ... }

}

A trait for a target architecture to be compiled to.

Required Methods

fn name(&self) -> &str

The name of the target architecture.

fn version(&self) -> &str

The version of the target architecture.

fn supports_floats(&self) -> bool

Whether or not the target architecture supports floating point.

fn supports_input(&self, src: &Input) -> bool

Whether or not the target architecture supports the given input (mode + channel).

fn supports_output(&self, dst: &Output) -> bool

Whether or not the target architecture supports the given output (mode + channel).

fn get(&mut self, src: &Input) -> Result<String, String>

Get a value from the given input stream (mode + channel).

fn put(&mut self, dst: &Output) -> Result<String, String>

Put a value to the given output stream (mode + channel).

fn peek(&mut self) -> Result<String, String>

Peek a value from the device connected to the program.

fn poke(&mut self) -> Result<String, String>

Poke a value to the device connected to the program.

fn declare_proc(&mut self, label_id: usize) -> String

Compile the declaration of a procedure.

fn end(&mut self, matching: &CoreOp, fun: Option<usize>) -> String

Compile an End instruction (with the matching If or While or Function)

fn op(&mut self, op: &CoreOp) -> String

Compile a CoreOp instruction.

fn std_op(&mut self, op: &StandardOp) -> Result<String, String>

Compile a StandardOp instruction.

Provided Methods

fn prelude(&self, _is_core: bool) -> Option<String>

The code before the program starts.

fn postop(&self) -> Option<String>

The code after each instruction.

fn postlude(&self, _is_core: bool) -> Option<String>

The code after the program ends.

fn pre_funs(&self, _funs: Vec<i32>) -> Option<String>

The code before the function definitions.

fn post_funs(&self, _funs: Vec<i32>) -> Option<String>

The code after the function definitions.

fn indentation(&self) -> Option<String>

The string used for indentation.