sage::targets
Trait CompiledTarget

pub trait CompiledTarget: Architecture {
    // Provided methods
    fn build_op(
        &mut self,
        op: &CoreOp,
        matching_ops: &mut Vec<CoreOp>,
        matching_funs: &mut Vec<usize>,
        current_fun: &mut usize,
        indent: &mut usize,
    ) -> Result<String, String> { ... }
    fn build_std_op(
        &mut self,
        std_op: &StandardOp,
        matching_ops: &mut Vec<CoreOp>,
        matching_funs: &mut Vec<usize>,
        current_fun: &mut usize,
        indent: &mut usize,
    ) -> Result<String, String> { ... }
    fn build_core(&mut self, program: &CoreProgram) -> Result<String, String> { ... }
    fn build_std(&mut self, program: &StandardProgram) -> Result<String, String> { ... }
}

Implement a compiler for the given target.

Provided Methods

fn build_op( &mut self, op: &CoreOp, matching_ops: &mut Vec<CoreOp>, matching_funs: &mut Vec<usize>, current_fun: &mut usize, indent: &mut usize, ) -> Result<String, String>

fn build_std_op( &mut self, std_op: &StandardOp, matching_ops: &mut Vec<CoreOp>, matching_funs: &mut Vec<usize>, current_fun: &mut usize, indent: &mut usize, ) -> Result<String, String>

fn build_core(&mut self, program: &CoreProgram) -> Result<String, String>

Compile the core variant of the machine code (must be implemented for every target).

fn build_std(&mut self, program: &StandardProgram) -> Result<String, String>

Compile the standard variant of the machine code (should be implemented for every target possible).