sage::vm
Struct StandardProgram

pub struct StandardProgram(pub Vec<StandardOp>);

A program of core and standard virtual machine instructions.

Tuple Fields
0: Vec<StandardOp>
Implementations

impl StandardProgram

pub fn flatten(self) -> Self

Flatten a core program so that all of its functions are defined sequentially at the beginning.

pub fn get_main(&self) -> Vec<StandardOp>

Get the code outside of any functions.

pub fn get_functions(&self) -> HashMap<i32, Vec<StandardOp>>

Get the code for each function.

pub fn get_main_and_functions( self, ) -> (Vec<StandardOp>, HashMap<i32, Vec<StandardOp>>)

Get the code outside of any functions, and the code for each function.