Struct sage::vm::TestingDevice

pub struct TestingDevice {
    pub ffi: HashMap<FFIBinding, fn(_: &mut VecDeque<i64>, _: Option<&mut Vec<i64>>)>,
    pub ffi_channel: VecDeque<i64>,
    pub input: VecDeque<i64>,
    pub output: Vec<(i64, Output)>,
}

A device used for testing the compiler. This simply keeps a buffer of sample input to supply to the virtual machine, and keeps an output buffer to keep track of the output of the virtual machine.

The tests interpret the program and populate the device with output. Then, we check the devices output against the correct output.

Fields
ffi: HashMap<FFIBinding, fn(_: &mut VecDeque<i64>, _: Option<&mut Vec<i64>>)>ffi_channel: VecDeque<i64>input: VecDeque<i64>output: Vec<(i64, Output)>
Implementations

impl TestingDevice

pub fn new(sample_input: impl ToString) -> Self

Create a new testing device with some given sample input.

pub fn new_raw(input: Vec<i64>) -> Self

pub fn output_str(&self) -> String

Get the output of the testing device as a string (ascii).

pub fn output_vals(&self) -> Vec<i64>