Module sage::asm

Assembly Module

This module contains a small assembly language implemented over the virtual machine. The assembly language, much like the virtual machine, is split into two variants: the “core” variant and the “standard” variant. The assembly language is not typed, and simply works by performing basic operations on one or more parameter addresses, and automatically managing procedure pointers. It also gives the programmer the ability to write inline assembly code for optimized performance or code size.

Index

Core Assembly

Standard Assembly

Assembly Memory Model

Global Variable Management

The Core Variant

The core variant is intended to be used with the core variant of the virtual machine. It is very portable, and MUST be supported on all implementations.

The Standard Variant

The standard variant is intended to be used with the standard variant of the virtual machine. It is very portable: it only adds instructions for float operations, memory allocation, and I/O.

Re-exports

pub use self::core::CoreOp;

pub use self::core::CoreProgram;

pub use self::std::StandardOp;

pub use self::std::StandardProgram;

pub use globals::Globals;

pub use location::Location;

pub use location::A;

pub use location::B;

pub use location::C;

pub use location::D;

pub use location::E;

pub use location::F;

pub use location::FP;

pub use location::GP;

pub use location::REGISTERS;

pub use location::SP;

Modules

core

Core Assembly Variant

globals

Assembly Global Variables

location

Assembly Memory Location

std

Standard Assembly Variant

Enums

Error

An error generated by assembling some assembly language code.

Traits

AssemblyProgram

A frontend to both the CoreProgram and StandardProgram types. This allows the compiler to append CoreOps to both programs with guaranteed success, but optionally allows the compiler to attempt to compile a StandardOp, which may fail depending on the target’s support for the instruction.