sage::asm::core
Enum CoreOp

pub enum CoreOp {

    Comment(String),
    Many(Vec<CoreOp>),
    Global {
        name: String,
        size: usize,
    },
    Set(Location, i64),
    VecSet(Location, Vec<i64>),
    SetLabel(Location, String),
    GetAddress {
        addr: Location,
        dst: Location,
    },
    PushAddress(Location),
    Call(Location),
    CallLabel(String),
    Return,
    Fn(String),
    While(Location),
    If(Location),
    Else,
    End,
    Move {
        src: Location,
        dst: Location,
    },
    Copy {
        src: Location,
        dst: Location,
        size: usize,
    },
    Swap(Location, Location),
    Next(Location, Option<isize>),
    Prev(Location, Option<isize>),
    Index {
        src: Location,
        offset: Location,
        dst: Location,
    },
    Inc(Location),
    Dec(Location),
    Add {
        src: Location,
        dst: Location,
    },
    Sub {
        src: Location,
        dst: Location,
    },
    Mul {
        src: Location,
        dst: Location,
    },
    Div {
        src: Location,
        dst: Location,
    },
    Rem {
        src: Location,
        dst: Location,
    },
    DivRem {
        src: Location,
        dst: Location,
    },
    Neg(Location),
    VecAdd {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecSub {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecMul {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecDiv {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecRem {
        size: usize,
        src: Location,
        dst: Location,
    },
    Not(Location),
    And {
        src: Location,
        dst: Location,
    },
    Or {
        src: Location,
        dst: Location,
    },
    VecNot {
        size: usize,
        dst: Location,
    },
    VecAnd {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecOr {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecInc {
        size: usize,
        dst: Location,
    },
    VecDec {
        size: usize,
        dst: Location,
    },
    LeftShift {
        src: Location,
        dst: Location,
    },
    LogicalRightShift {
        src: Location,
        dst: Location,
    },
    ArithmeticRightShift {
        src: Location,
        dst: Location,
    },
    VecLeftShift {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecLogicalRightShift {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecArithmeticRightShift {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecGez {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecNeg {
        size: usize,
        dst: Location,
    },
    VecOffset {
        size: usize,
        dst: Location,
        offset: isize,
    },
    VecIndex {
        size: usize,
        src: Location,
        offset: Location,
        dst: Location,
    },
    Push(Location, usize),
    Pop(Option<Location>, usize),
    PushTo {
        src: Location,
        sp: Location,
        size: usize,
    },
    PopFrom {
        sp: Location,
        dst: Option<Location>,
        size: usize,
    },
    Compare {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsGreater {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsGreaterEqual {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsLess {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsLessEqual {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsEqual {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsNotEqual {
        a: Location,
        b: Location,
        dst: Location,
    },
    Get(Location, Input),
    Put(Location, Output),
    Array {
        src: Location,
        dst: Location,
        vals: Vec<i64>,
    },
    Const {
        dst: Location,
        vals: Vec<i64>,
    },
    PushConst(Vec<i64>),
    BitwiseNand {
        src: Location,
        dst: Location,
    },
    BitwiseXor {
        src: Location,
        dst: Location,
    },
    BitwiseOr {
        src: Location,
        dst: Location,
    },
    BitwiseNor {
        src: Location,
        dst: Location,
    },
    BitwiseAnd {
        src: Location,
        dst: Location,
    },
    BitwiseNot(Location),
    VecBitwiseAnd {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecBitwiseOr {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecBitwiseXor {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecBitwiseNor {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecBitwiseNand {
        size: usize,
        src: Location,
        dst: Location,
    },
    VecBitwiseNot {
        size: usize,
        dst: Location,
    },

}

A core instruction of the assembly language. These are instructions guaranteed to be implemented for every target possible.

Variants

Comment(String)

Many(Vec<CoreOp>)

Many instructions to execute; conveniently grouped together. This is useful for code generation.

Global

Declare a global variable.

This will allow the program to reference the global variable using $name in the assembly code, where name is the name of the global variable.

To access the first element of the global variable, use $name. To access the second element of the global variable, use $name + 1.

To get the address of the first element of the global variable, use lea $name, $tmp.

Fields

name: String

size: usize

Set(Location, i64)

Set the value of a register, or any location in memory, to a given constant.

VecSet(Location, Vec<i64>)

Set the vector values of a destination.

SetLabel(Location, String)

Set the value of a register, or any location in memory, to the value of a label’s ID.

GetAddress

Get the address of a location, and store it in a destination

Fields

addr: Location

dst: Location

PushAddress(Location)

Get the address of a location and push it to the stack.

Call(Location)

Get a value in memory and call it as a label ID.

CallLabel(String)

Call a function with a given label.

Return

Return from the current function.

Fn(String)

Declare a new label.

While(Location)

Begin a “while the value is not zero” loop over a given register or location in memory.

If(Location)

Begin an “if the value is not zero” statement over a given register or location in memory.

Else

Add an “else” clause to an “if the value is not zero” statement.

End

Terminate a function declaration, a while loop, an if statement, or an else clause.

Move

Copy a value from a source location to a destination location.

dst = src

Fields

src: Location

dst: Location

Copy

Copy a number of cells from a source referenced location to a destination referenced location.

This will dereference src, copy its contents, and store them at the location pointed to by dst.

*dst = *src

Fields

src: Location

dst: Location

size: usize

Swap(Location, Location)

Swap the values of two locations.

Next(Location, Option<isize>)

Make this pointer point to the next cell (or the nth next cell).

Prev(Location, Option<isize>)

Make this pointer point to the previous cell (or the nth previous cell).

Index

Get the address of a location indexed by an offset stored at another location. Store the result in the destination register.

This is essentially the runtime equivalent of dst = addr.offset(offset)

Fields

src: Location

offset: Location

dst: Location

Inc(Location)

Increment the integer value of a location.

Dec(Location)

Decrement the integer value of a location.

Add

Add an integer value from a source location to a destination location.

Fields

src: Location

dst: Location

Sub

Subtract a source integer value from a destination location.

Fields

src: Location

dst: Location

Mul

Multiply a destination location by a source value.

Fields

src: Location

dst: Location

Div

Divide a destination location by a source value.

Fields

src: Location

dst: Location

Rem

Store the remainder of the destination modulus the source in the destination.

Fields

src: Location

dst: Location

DivRem

Divide a destination location by a source value. Store the quotient in the destination, and the remainder in the source.

Fields

src: Location

dst: Location

Neg(Location)

Negate an integer.

VecAdd

Perform a SIMD addition over a vector of integers. This will add the source vector to the destination vector.

Fields

size: usize

src: Location

dst: Location

VecSub

Perform a SIMD subtraction over a vector of integers. This will subtract the source vector from the destination vector.

Fields

size: usize

src: Location

dst: Location

VecMul

Perform a SIMD multiplication over a vector of integers. This will multiply the source vector with the destination vector.

Fields

size: usize

src: Location

dst: Location

VecDiv

Perform a SIMD division over a vector of integers. This will divide the destination vector by the source vector.

Fields

size: usize

src: Location

dst: Location

VecRem

Perform a SIMD remainder over a vector of integers. This will store the remainder of the destination vector divided by the source vector.

Fields

size: usize

src: Location

dst: Location

Not(Location)

Replace a value in memory with its boolean complement.

And

Logical “and” a destination with a source value.

Fields

src: Location

dst: Location

Or

Logical “or” a destination with a source value.

Fields

src: Location

dst: Location

VecNot

Perform a vector “Not” operation. This will replace every value in the vector with its boolean complement.

Fields

size: usize

dst: Location

VecAnd

Perform a vector “And” operation. This will perform a logical “and” operation on every value in the vectors.

Fields

size: usize

src: Location

dst: Location

VecOr

Perform a vector “Or” operation. This will perform a logical “or” operation on every value in the vectors.

Fields

size: usize

src: Location

dst: Location

VecInc

Increment a vector of integers.

Fields

size: usize

dst: Location

VecDec

Decrement a vector of integers.

Fields

size: usize

dst: Location

LeftShift

Left shift a destination by a source value.

Fields

src: Location

dst: Location

LogicalRightShift

Logical right shift a destination by a source value. This will fill the leftmost bits with zeroes.

Fields

src: Location

dst: Location

ArithmeticRightShift

Arithmetic right shift a destination by a source value. This will fill the leftmost bits with the sign bit.

Fields

src: Location

dst: Location

VecLeftShift

Perform a SIMD left shift operation over a vector of integers.

Fields

size: usize

src: Location

dst: Location

VecLogicalRightShift

Perform a SIMD logical right shift operation over a vector of integers.

Fields

size: usize

src: Location

dst: Location

VecArithmeticRightShift

Perform a SIMD arithmetic right shift operation over a vector of integers.

Fields

size: usize

src: Location

dst: Location

VecGez

Perform a SIMD “is >= zero” operation over a vector of integers. Store the result in the destination vector.

Fields

size: usize

src: Location

dst: Location

VecNeg

Negate a vector of integers using SIMD.

Fields

size: usize

dst: Location

VecOffset

Perform a SIMD pointer arithmetic operation over a vector of integers. This will add the “offset” to every value in the vector, so that each pointer in the vector is offset by the same amount.

Fields

size: usize

dst: Location

offset: isize

VecIndex

Perform a SIMD pointer index operation over a vector of integers. This will index every pointer in the vector by the integers in the source vector.

Fields

size: usize

src: Location

offset: Location

dst: Location

Push(Location, usize)

Push a number of cells starting at a memory location on the stack.

Pop(Option<Location>, usize)

Pop a number of cells from the stack and store it in a memory location.

PushTo

Push a number of cells starting at a memory location onto a specified stack. This will increment the stack’s pointer by the number of cells pushed.

Fields

src: Location

sp: Location

size: usize

PopFrom

Pop a number of cells from a specified stack and store it in a memory location. This will decrement the stack’s pointer by the number of cells popped.

Fields

sp: Location

dst: Option<Location>

size: usize

Compare

Store the comparison of “a” and “b” in a destination register. If “a” is less than “b”, store -1. If “a” is greater than “b”, store 1. If “a” is equal to “b”, store 0.

Fields

a: Location

b: Location

dst: Location

IsGreater

Perform dst = a > b.

Fields

a: Location

b: Location

dst: Location

IsGreaterEqual

Perform dst = a >= b.

Fields

a: Location

b: Location

dst: Location

IsLess

Perform dst = a < b.

Fields

a: Location

b: Location

dst: Location

IsLessEqual

Perform dst = a <= b.

Fields

a: Location

b: Location

dst: Location

IsEqual

Perform dst = a == b.

Fields

a: Location

b: Location

dst: Location

IsNotEqual

Perform dst = a != b.

Fields

a: Location

b: Location

dst: Location

Get(Location, Input)

Get a value from the input device / interface and store it in a destination register.

Put(Location, Output)

Put a value from a source register to the output device / interface.

Array

Store a list of values at a source location. Then, store the address past the last value into the destination location.

Fields

src: Location

dst: Location

vals: Vec<i64>

Const

Write a set of values to a memory location.

Fields

dst: Location

vals: Vec<i64>

PushConst(Vec<i64>)

Push a const to the stack.

BitwiseNand

Fields

src: Location

dst: Location

BitwiseXor

Fields

src: Location

dst: Location

BitwiseOr

Fields

src: Location

dst: Location

BitwiseNor

Fields

src: Location

dst: Location

BitwiseAnd

Fields

src: Location

dst: Location

BitwiseNot(Location)

VecBitwiseAnd

Fields

size: usize

src: Location

dst: Location

VecBitwiseOr

Fields

size: usize

src: Location

dst: Location

VecBitwiseXor

Fields

size: usize

src: Location

dst: Location

VecBitwiseNor

Fields

size: usize

src: Location

dst: Location

VecBitwiseNand

Fields

size: usize

src: Location

dst: Location

VecBitwiseNot

Fields

size: usize

dst: Location

Implementations

impl CoreOp

pub fn put_string(msg: impl ToString, dst: Output) -> Self

Put a string literal as UTF-8 to the output device.

pub fn push_string(msg: impl ToString) -> Self

Push a string literal as UTF-8 to the stack.

pub fn stack_alloc_string(dst: Location, text: impl ToString) -> Self

Allocate a string on the stack, and store its address in a destination register.