Enum sage::asm::core::CoreOp

pub enum CoreOp {

    Comment(String),
    Many(Vec<CoreOp>),
    Set(Location, i64),
    SetLabel(Location, String),
    GetAddress {
        addr: Location,
        dst: Location,
    },
    Call(Location),
    CallLabel(String),
    Return,
    Fn(String),
    While(Location),
    If(Location),
    Else,
    End,
    Move {
        src: Location,
        dst: Location,
    },
    Copy {
        src: Location,
        dst: Location,
        size: usize,
    },
    Swap(Location, Location),
    Next(Location, Option<isize>),
    Prev(Location, Option<isize>),
    Index {
        src: Location,
        offset: Location,
        dst: Location,
    },
    Inc(Location),
    Dec(Location),
    Add {
        src: Location,
        dst: Location,
    },
    Sub {
        src: Location,
        dst: Location,
    },
    Mul {
        src: Location,
        dst: Location,
    },
    Div {
        src: Location,
        dst: Location,
    },
    Rem {
        src: Location,
        dst: Location,
    },
    DivRem {
        src: Location,
        dst: Location,
    },
    Neg(Location),
    Not(Location),
    And {
        src: Location,
        dst: Location,
    },
    Or {
        src: Location,
        dst: Location,
    },
    Push(Location, usize),
    Pop(Option<Location>, usize),
    PushTo {
        src: Location,
        sp: Location,
        size: usize,
    },
    PopFrom {
        sp: Location,
        dst: Option<Location>,
        size: usize,
    },
    Compare {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsGreater {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsGreaterEqual {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsLess {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsLessEqual {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsEqual {
        a: Location,
        b: Location,
        dst: Location,
    },
    IsNotEqual {
        a: Location,
        b: Location,
        dst: Location,
    },
    Get(Location, Input),
    Put(Location, Output),
    Array {
        src: Location,
        dst: Location,
        vals: Vec<i64>,
    },
    BitwiseNand {
        src: Location,
        dst: Location,
    },
    BitwiseXor {
        src: Location,
        dst: Location,
    },
    BitwiseOr {
        src: Location,
        dst: Location,
    },
    BitwiseNor {
        src: Location,
        dst: Location,
    },
    BitwiseAnd {
        src: Location,
        dst: Location,
    },
    BitwiseNot(Location),

}

A core instruction of the assembly language. These are instructions guaranteed to be implemented for every target possible.

Variants

Comment(String)

Many(Vec<CoreOp>)

Set(Location, i64)

Set the value of a register, or any location in memory, to a given constant.

SetLabel(Location, String)

Set the value of a register, or any location in memory, to the value of a label’s ID.

GetAddress

Fields

addr: Location

dst: Location

Get the address of a location, and store it in a destination

Call(Location)

Get a value in memory and call it as a label ID.

CallLabel(String)

Call a function with a given label.

Return

Return from the current function.

Fn(String)

Declare a new label.

While(Location)

Begin a “while the value is not zero” loop over a given register or location in memory.

If(Location)

Begin an “if the value is not zero” statement over a given register or location in memory.

Else

Add an “else” clause to an “if the value is not zero” statement.

End

Terminate a function declaration, a while loop, an if statement, or an else clause.

Move

Fields

src: Location

dst: Location

Copy a value from a source location to a destination location.

dst = src

Copy

Fields

src: Location

dst: Location

size: usize

Copy a number of cells from a source referenced location to a destination referenced location.

This will dereference src, copy its contents, and store them at the location pointed to by dst.

*dst = *src

Swap(Location, Location)

Swap the values of two locations.

Next(Location, Option<isize>)

Make this pointer point to the next cell (or the nth next cell).

Prev(Location, Option<isize>)

Make this pointer point to the previous cell (or the nth previous cell).

Index

Fields

src: Location

offset: Location

dst: Location

Get the address of a location indexed by an offset stored at another location. Store the result in the destination register.

This is essentially the runtime equivalent of dst = addr.offset(offset)

Inc(Location)

Increment the integer value of a location.

Dec(Location)

Decrement the integer value of a location.

Add

Fields

src: Location

dst: Location

Add an integer value from a source location to a destination location.

Sub

Fields

src: Location

dst: Location

Subtract a source integer value from a destination location.

Mul

Fields

src: Location

dst: Location

Multiply a destination location by a source value.

Div

Fields

src: Location

dst: Location

Divide a destination location by a source value.

Rem

Fields

src: Location

dst: Location

Store the remainder of the destination modulus the source in the destination.

DivRem

Fields

src: Location

dst: Location

Divide a destination location by a source value. Store the quotient in the destination, and the remainder in the source.

Neg(Location)

Negate an integer.

Not(Location)

Replace a value in memory with its boolean complement.

And

Fields

src: Location

dst: Location

Logical “and” a destination with a source value.

Or

Fields

src: Location

dst: Location

Logical “or” a destination with a source value.

Push(Location, usize)

Push a number of cells starting at a memory location on the stack.

Pop(Option<Location>, usize)

Pop a number of cells from the stack and store it in a memory location.

PushTo

Fields

src: Location

sp: Location

size: usize

Push a number of cells starting at a memory location onto a specified stack. This will increment the stack’s pointer by the number of cells pushed.

PopFrom

Fields

sp: Location

dst: Option<Location>

size: usize

Pop a number of cells from a specified stack and store it in a memory location. This will decrement the stack’s pointer by the number of cells popped.

Compare

Fields

a: Location

b: Location

dst: Location

Store the comparison of “a” and “b” in a destination register. If “a” is less than “b”, store -1. If “a” is greater than “b”, store 1. If “a” is equal to “b”, store 0.

IsGreater

Fields

a: Location

b: Location

dst: Location

Perform dst = a > b.

IsGreaterEqual

Fields

a: Location

b: Location

dst: Location

Perform dst = a >= b.

IsLess

Fields

a: Location

b: Location

dst: Location

Perform dst = a < b.

IsLessEqual

Fields

a: Location

b: Location

dst: Location

Perform dst = a <= b.

IsEqual

Fields

a: Location

b: Location

dst: Location

Perform dst = a == b.

IsNotEqual

Fields

a: Location

b: Location

dst: Location

Perform dst = a != b.

Get(Location, Input)

Get a value from the input device / interface and store it in a destination register.

Put(Location, Output)

Put a value from a source register to the output device / interface.

Array

Fields

src: Location

dst: Location

vals: Vec<i64>

Store a list of values at a source location. Then, store the address past the last value into the destination location.

BitwiseNand

Fields

src: Location

dst: Location

BitwiseXor

Fields

src: Location

dst: Location

BitwiseOr

Fields

src: Location

dst: Location

BitwiseNor

Fields

src: Location

dst: Location

BitwiseAnd

Fields

src: Location

dst: Location

BitwiseNot(Location)

Implementations

impl CoreOp

pub fn put_string(msg: impl ToString, dst: Output) -> Self

Put a string literal as UTF-8 to the output device.

pub fn push_string(msg: impl ToString) -> Self

pub fn stack_alloc_cells(dst: Location, vals: Vec<i64>) -> Self

pub fn stack_alloc_string(dst: Location, text: impl ToString) -> Self