Trait sage::asm::AssemblyProgram

pub trait AssemblyProgram {
    // Required methods
    fn op(&mut self, op: CoreOp);
    fn std_op(&mut self, op: StandardOp) -> Result<(), Error>;

    // Provided method
    fn comment(&mut self, comment: String) { ... }
}

A frontend to both the CoreProgram and StandardProgram types. This allows the compiler to append CoreOps to both programs with guaranteed success, but optionally allows the compiler to attempt to compile a StandardOp, which may fail depending on the target’s support for the instruction.

Required Methods

fn op(&mut self, op: CoreOp)

fn std_op(&mut self, op: StandardOp) -> Result<(), Error>

Provided Methods

fn comment(&mut self, comment: String)