Enum sage::io::OutputMode

pub enum OutputMode {

    StdoutChar,
    StdoutInt,
    StdoutFloat,
    StderrChar,
    StderrInt,
    StderrFloat,
    PrinterChar,
    PrinterInt,
    PrinterFloat,
    Brightness,
    AnalogPin,
    DigitalPin,
    StepperMotor,
    Solenoid,
    Valve,
    MotorSpeed,
    Servo,
    Temperature,
    Pump,
    Fan,
    Blower,
    Heater,
    Cooler,
    Pressure,
    Buzzer,
    Bell,
    Note,
    SpeakerVolume,
    SpeakerFrequency,
    UpdateDisplay,
    ClearDisplay,
    SetCursorRow,
    SetCursorColumn,
    MoveCursorUp,
    MoveCursorDown,
    MoveCursorLeft,
    MoveCursorRight,
    SetCursorChar(Color),
    SetCursorPixel(Color),
    Custom(String),

}

The different types of output modes a program might use.

Variants

StdoutChar

Standard output modes Standard output (ASCII character)

StdoutInt

Standard output (integer)

StdoutFloat

Standard output (float)

StderrChar

Standard error (ASCII character)

StderrInt

Standard error (integer)

StderrFloat

Standard error (float)

PrinterChar

Alternative output modes for standard output Printer (ASCII character)

PrinterInt

Printer (integer)

PrinterFloat

Printer (float)

Brightness

Lighting device output modes Set the brightness of a given light (in percent)

AnalogPin

Electrical device output modes Set the voltage of a given analog output

DigitalPin

Set the state of a given digital output (0=low, 1=high)

StepperMotor

Robotics device output modes Set the position of a given stepper motor (in radians)

Solenoid

Set the polarity of a solenoid (0=off, 1=on)

Valve

Set the position of a given valve (0=closed, 1=open)

MotorSpeed

Set the speed of a given motor (in revolutions per minute)

Servo

Set the position of a given servo (in radians)

Temperature

Set the temperature of a given heating/cooling device (degrees K)

Pump

Set the pressure of a given pump (in atmospheres)

Fan

Turn a fan on or off (0=off, 1=on)

Blower

Set the pressure of a given blower (in atmospheres)

Heater

Turn a heater on or off (0=off, 1=on)

Cooler

Turn a cooler on or off (0=off, 1=on)

Pressure

Set the pressure of a given vacuum/pressurizer (atmospheres)

Buzzer

Sound output modes Ring a given buzzer (in hertz)

Bell

Ring a bell (in hertz)

Note

Play a given note (in hertz)

SpeakerVolume

Set the volume of a given speaker (in percent)

SpeakerFrequency

Set the frequency of a given speaker (in hertz)

UpdateDisplay

Display output modes Update the display

ClearDisplay

Clear the display

SetCursorRow

Set the cursor row on the display

SetCursorColumn

Set the cursor column on the display

MoveCursorUp

Move the cursor up on the display

MoveCursorDown

Move the cursor down on the display

MoveCursorLeft

Move the cursor left on the display

MoveCursorRight

Move the cursor right on the display

SetCursorChar(Color)

Write a character to the display

SetCursorPixel(Color)

Set the color of a given pixel on the display

Custom(String)

Custom output modes A custom output mode (for use with a custom output device)