Enum sage::lir::Expr

pub enum Expr {

    AnnotatedWithSource {
        expr: Box<Self>,
        loc: SourceCodeLocation,
    },
    ConstExpr(ConstExpr),
    Many(Vec<Self>),
    LetConst(String, ConstExpr, Box<Self>),
    LetConsts(BTreeMap<String, ConstExpr>, Box<Self>),
    LetProc(String, Procedure, Box<Self>),
    LetProcs(Vec<(String, Procedure)>, Box<Self>),
    LetType(String, Type, Box<Self>),
    LetTypes(Vec<(String, Type)>, Box<Self>),
    LetVar(String, Option<Type>, Box<Self>, Box<Self>),
    LetVars(Vec<(String, Option<Type>, Self)>, Box<Self>),
    While(Box<Self>, Box<Self>),
    If(Box<Self>, Box<Self>, Box<Self>),
    When(ConstExpr, Box<Self>, Box<Self>),
    Match(Box<Self>, Vec<(Pattern, Self)>),
    IfLet(Pattern, Box<Self>, Box<Self>, Box<Self>),
    UnaryOp(Box<dyn UnaryOp>, Box<Self>),
    BinaryOp(Box<dyn BinaryOp>, Box<Self>, Box<Self>),
    TernaryOp(Box<dyn TernaryOp>, Box<Self>, Box<Self>, Box<Self>),
    AssignOp(Box<dyn AssignOp>, Box<Self>, Box<Self>),
    Refer(Box<Self>),
    Deref(Box<Self>),
    DerefMut(Box<Self>, Box<Self>),
    Apply(Box<Self>, Vec<Self>),
    Return(Box<Self>),
    Array(Vec<Self>),
    Tuple(Vec<Self>),
    Union(Type, String, Box<Self>),
    EnumUnion(Type, String, Box<Self>),
    Struct(BTreeMap<String, Self>),
    As(Box<Self>, Type),
    Member(Box<Self>, ConstExpr),
    Index(Box<Self>, Box<Self>),

}

TODO: Add variants for LetProc, LetVar, etc. to support multiple definitions. This way, we don’t overflow the stack with several clones of the environment. A runtime expression.

Variants

AnnotatedWithSource

Fields

expr: Box<Self>

loc: SourceCodeLocation

The source code location of the expression.

An expression along with data about its source code location. This is used for error reporting.

ConstExpr(ConstExpr)

A constant expression.

Many(Vec<Self>)

A block of expressions. The last expression in the block is the value of the block.

LetConst(String, ConstExpr, Box<Self>)

A const binding expression. Declare a constant under a new scope, and evaluate a subexpression in that scope.

LetConsts(BTreeMap<String, ConstExpr>, Box<Self>)

A const binding expression. Declare multiple constants under a new scope, and evaluate a subexpression in that scope.

LetProc(String, Procedure, Box<Self>)

A proc binding expression. Declare a procedure under a new scope, and evaluate a subexpression in that scope.

LetProcs(Vec<(String, Procedure)>, Box<Self>)

A proc binding expression. Declare multiple procedures under a new scope, and evaluate a subexpression in that scope.

LetType(String, Type, Box<Self>)

A type binding expression. Declare a type under a new scope, and evaluate a subexpression in that scope.

LetTypes(Vec<(String, Type)>, Box<Self>)

A type binding expression. Declare multiple types under a new scope, and evaluate a subexpression in that scope.

LetVar(String, Option<Type>, Box<Self>, Box<Self>)

A let binding expression. Declare a variable under a new scope, and evaluate a subexpression in that scope.

LetVars(Vec<(String, Option<Type>, Self)>, Box<Self>)

A let binding expression. Declare multiple variables under a new scope, and evaluate a subexpression in that scope.

While(Box<Self>, Box<Self>)

Create a while loop: while the first expression evaluates to true, evaluate the second expression.

If(Box<Self>, Box<Self>, Box<Self>)

An if-then-else expression.

Evaluate a condition. If the condition is true, evaluate the first expression. Otherwise, evaluate the second expression.

When(ConstExpr, Box<Self>, Box<Self>)

A constant, compile time if-then-else expression.

Evaluate the condition as a constant expression. If the condition is true, then this when expression is replaced with the first expression. Otherwise, this when expression is replaced with the second expression.

Match(Box<Self>, Vec<(Pattern, Self)>)

A match expression.

IfLet(Pattern, Box<Self>, Box<Self>, Box<Self>)

An if-let expression.

This attempts to match an expression against a pattern, and if it succeeds, evaluates the body expression with the matched variables in scope. If the match fails, the else expression is evaluated.

UnaryOp(Box<dyn UnaryOp>, Box<Self>)

Perform a unary operation on two expressions.

BinaryOp(Box<dyn BinaryOp>, Box<Self>, Box<Self>)

Perform a binary operation on two expressions.

TernaryOp(Box<dyn TernaryOp>, Box<Self>, Box<Self>, Box<Self>)

Perform a ternary operation on three expressions.

AssignOp(Box<dyn AssignOp>, Box<Self>, Box<Self>)

Perform an assignment operation on two expressions.

Refer(Box<Self>)

Reference this expression (i.e. get a pointer to it).

Deref(Box<Self>)

Dereference this expression (i.e. get the value it points to).

DerefMut(Box<Self>, Box<Self>)

Store an expression to an address (a pointer).

Apply(Box<Self>, Vec<Self>)

Apply a function with some arguments.

Return(Box<Self>)

Return a value from a function.

Array(Vec<Self>)

An array of expressions.

Tuple(Vec<Self>)

A tuple of expressions.

Union(Type, String, Box<Self>)

A union: a collection of named fields. The Type value is the type of the union. The String field is the field the union is being initialized with. The Box<Self> field is the value of the field we want to initialize.

EnumUnion(Type, String, Box<Self>)

A tagged union: a typechecked union of different variants. The Type value is the type of the tagged union. The String field is the variant the tagged union is being initialized with. The Box<Self> field is the value of the union’s data we want to initialize.

Struct(BTreeMap<String, Self>)

A structure of fields to expressions.

As(Box<Self>, Type)

Cast an expression to another type.

Member(Box<Self>, ConstExpr)

Get a field or member from a structure, union, or tuple. For tuples, use an Int constant expression to access the nth field (zero indexed). For unions or structures, use a Symbol constant expression to access the field.

Index(Box<Self>, Box<Self>)

Index an array or pointer with an expression that evaluates to an Int at runtime.

Implementations

impl Expr

pub fn size_of(self) -> Self

Get the size of an expression.

pub fn as_type(self, t: Type) -> Self

Cast an expression as another type.

pub fn unop(self, op: impl UnaryOp + 'static) -> Self

pub fn not(self) -> Self

Logical not this expression.

pub fn bitxor(self, other: impl Into<Self>) -> Self

Bitwise this expression with another.

pub fn bitor(self, other: impl Into<Self>) -> Self

BitwiseOr this expression with another.

pub fn bitnand(self, other: impl Into<Self>) -> Self

BitwiseOr this expression with another.

pub fn bitand(self, other: impl Into<Self>) -> Self

BitwiseAnd this expression with another.

pub fn bitnor(self, other: impl Into<Self>) -> Self

BitwiseAnd this expression with another.

pub fn bitnot(self) -> Self

BitwiseAnd this expression with another.

pub fn lt(self, other: impl Into<Self>) -> Self

Is this expression less than another?

pub fn le(self, other: impl Into<Self>) -> Self

Is this expression less than or equal to another?

pub fn gt(self, other: impl Into<Self>) -> Self

Is this expression greater than another?

pub fn ge(self, other: impl Into<Self>) -> Self

Is this expression greater than or equal to another?

pub fn eq(self, other: impl Into<Self>) -> Self

Is this expression greater than another?

pub fn neq(self, other: impl Into<Self>) -> Self

Is this expression greater than or equal to another?

pub fn and(self, other: impl Into<Self>) -> Self

Logical and this expression with another.

pub fn or(self, other: impl Into<Self>) -> Self

Logical or this expression with another.

pub fn pow(self, exp: impl Into<Self>) -> Self

Get the power of this expression to another.

pub fn add(self, other: impl Into<Self>) -> Self

Add this expression to another.

pub fn sub(self, other: impl Into<Self>) -> Self

Subtract an expression from this expression.

pub fn mul(self, other: impl Into<Self>) -> Self

Multiply this expression by another.

pub fn div(self, other: impl Into<Self>) -> Self

Divide this expression by another.

pub fn rem(self, other: impl Into<Self>) -> Self

Get the remainder of this expression divided by another.

pub fn neg(self) -> Self

Get the remainder of this expression divided by another.

pub fn field(self, field: ConstExpr) -> Self

Get a field from a structure, union, or tuple.

For tuples, use an Int constant expression to access the nth field (zero indexed). For unions or structures, use a Symbol constant expression to access the field.

pub fn idx(self, idx: impl Into<Self>) -> Self

Index an array or pointer with an expression that evaluates to an Int at runtime.

pub fn let_var( var: impl ToString, t: Option<Type>, e: impl Into<Self>, ret: impl Into<Self> ) -> Self

Create a let binding for an expression.

This will create a new scope with the variable var defined. var will be declared with the type t, and the expression e will be assigned to it.

The result of this expression is the ret value, which is evaluated under this new scope.

When this expression is finished evaluating, var will be removed from the scope.

pub fn let_vars( vars: Vec<(&str, Option<Type>, Self)>, ret: impl Into<Self> ) -> Self

Create a let binding for an expression, and define multiple variables.

pub fn let_type(typename: impl ToString, t: Type, ret: impl Into<Self>) -> Self

Create a let binding for an type.

This will create a new scope with the type typename defined. typename will be declared with the type t, and the expression ret will be evaluated under this new scope.

When this expression is finished evaluating, typename will be removed from the scope.

pub fn let_types(vars: Vec<(&str, Type)>, ret: impl Into<Self>) -> Self

Create several type bindings at onces.

pub fn let_const( constname: impl ToString, e: ConstExpr, ret: impl Into<Self> ) -> Self

Create a let binding for a constant expression.

This will create a new scope with the constant constname defined. ret will be evaluated under this new scope.

When this expression is finished evaluating, constname will be removed from the scope.

pub fn let_consts( constants: Vec<(&str, ConstExpr)>, ret: impl Into<Self> ) -> Self

Create several const bindings at onces.

pub fn let_proc( procname: impl ToString, proc: Procedure, ret: impl Into<Self> ) -> Self

Create a proc binding for a procedure.

This will create a new scope with the procedure proc defined. ret will be evaluated under this new scope.

When this expression is finished evaluating, proc will be removed from the scope.

pub fn let_procs(procs: BTreeMap<&str, Procedure>, ret: impl Into<Self>) -> Self

Create several proc bindings at onces.

pub fn structure(vars: BTreeMap<&str, Self>) -> Self

Create a structure of fields to expressions.

pub fn var(var: impl ToString) -> Self

Evaluate a variable in the current scope.

pub fn app(self, args: Vec<Self>) -> Self

Apply this expression as a procedure to some arguments.

pub fn if_then(self, t: impl Into<Self>, e: impl Into<Self>) -> Self

Create an if-then-else statement with this expression as the condition.

pub fn while_loop(self, body: impl Into<Self>) -> Self

Create a while statement with this expression as the condition.

pub fn refer(self) -> Self

Reference this expression (i.e. get a pointer to it).

pub fn deref(self) -> Self

Dereference this expression (i.e. get the value it points to).

pub fn deref_mut(self, e: impl Into<Self>) -> Self

Dereference this expression (i.e. get the value it points to), and write another expression to its position in memory.

pub fn assign_op(self, op: impl AssignOp + 'static, e: impl Into<Self>) -> Self

Perform an AssignOp on this expression.

pub fn assign(self, op: Box<dyn AssignOp>, e: impl Into<Self>) -> Self

Perform an AssignOp on this expression.