Enum sage::lir::Type

pub enum Type {

    Let(String, Box<Self>, Box<Self>),
    Unit(String, Box<Self>),
    Symbol(String),
    None,
    Int,
    Float,
    Cell,
    Char,
    Bool,
    Enum(Vec<String>),
    Tuple(Vec<Self>),
    Array(Box<Self>, Box<ConstExpr>),
    Struct(BTreeMap<String, Self>),
    EnumUnion(BTreeMap<String, Self>),
    Type(Box<Type>),
    Union(BTreeMap<String, Self>),
    Proc(Vec<Type>, Box<Type>),
    Pointer(Mutability, Box<Self>),
    Any,
    Never,
    Poly(Vec<String>, Box<Self>),
    Apply(Box<Self>, Vec<Self>),

}

The representation of a type in the LIR type system.

Variants

Let(String, Box<Self>, Box<Self>)

Bind a type to a name in a temporary scope.

Unit(String, Box<Self>)

This type is identified by its name. Most types are checked according to structural equality, but this type is checked according to name. Structural equality is also verified in addition to name equality.

The inner type acts exactly the same, but it can only be type-equal to another Unit type with the same name and inner type.

Symbol(String)

A named type.

None

The type of void expressions.

Int

The integer type.

Float

The floating-point number type.

Cell

The type of the most basic unit of memory.

Char

The type of a character.

Bool

The type of a boolean value.

Enum(Vec<String>)

An enumeration of a list of possible named values. A boolean could be considered an enumeration of true and false.

Tuple(Vec<Self>)

A heterogenous collection of types. This is a product type.

Array(Box<Self>, Box<ConstExpr>)

An array of a given type, with a constant size.

Struct(BTreeMap<String, Self>)

A tuple with named members. This is a product type.

EnumUnion(BTreeMap<String, Self>)

An enumeration of a list of possible types. This is a sum type. The enum union is essentially a regular union type, but with a tag to guarantee typesafety of accessing members.

The tag is stored at the very beginning of the value.

Type(Box<Type>)

A trait object. This is internally represented as an EnumUnion over the possible types that can be stored in the trait object. A type. This converts a type into a value.

Union(BTreeMap<String, Self>)

A union of a list of possible types mapped to named members. A union’s value is reinterpreted as a single type, depending on the member accessed. Unions’ values are stored starting at the beginning of the union’s address in memory, and are padded at the end with zeroes.

This is a sum type.

Proc(Vec<Type>, Box<Type>)

A procedure with a list of parameters and a return type.

Pointer(Mutability, Box<Self>)

A pointer to another type.

Any

A type reserved by the compiler. This type is equal to any other type. The NULL pointer, for example, is of type Pointer(Any).

Never

The type of an expression that will never return, or doesn’t resolve to a value.

Poly(Vec<String>, Box<Self>)

A polymorphic, parametric type. This type is used with the Apply type to create a concrete type. The type takes a list of symbols that are substituted with concrete types. The type is then simplified to a concrete type when combined with Apply. This type is used to implement generics.

Apply(Box<Self>, Vec<Self>)

A type that constructs a concrete type from a polymorphic type. This type is used to implement generics.

Implementations

impl Type

pub const SIMPLIFY_RECURSION_LIMIT: usize = 30usize

This is the maximum number of times a type will be simplified recursively.

pub fn is_recursive(&self, env: &Env) -> Result<bool, Error>

pub fn is_recursive_helper( &self, symbols: &mut HashSet<String>, env: &Env ) -> Result<bool, Error>

pub fn strip_template(&self, env: &Env) -> Self

pub fn is_poly(&self) -> bool

pub fn get_monomorph_template_args( &self, template: &Self, matched_symbols: &mut HashMap<String, Self>, param_symbols: &HashSet<String>, env: &Env ) -> Result<(), Error>

pub fn is_monomorph_of(&self, template: &Self, env: &Env) -> Result<bool, Error>

pub fn add_monomorphized_associated_consts( &self, env: &Env ) -> Result<(), Error>

pub fn get_template_params(&self, env: &Env) -> Vec<String>

pub fn apply(&self, args: Vec<Self>) -> Self

pub fn is_simple(&self) -> bool

Is this type in a simple form? A simple form is a form that does not require any immediate simplification. For example, a Poly type is not simple, because it requires a Apply type to deduce anything about the type. A let binding is also not simple, because it requires the inner type to be substituted for the name. A symbol is not simple, because it requires a lookup in the environment to determine the type. Other types, like a tagged union, which may be pattern matched over the inner types, does not require an immediate lookup to use some of its type information.

pub fn is_concrete(&self) -> bool

pub fn is_atomic(&self) -> bool

Is this type an irreducible, atomic type?

pub fn is_self_param_reference(&self, env: &Env) -> Result<bool, Error>

Is first argument of function a reference?

pub fn get_self_param_mutability(&self, env: &Env) -> Option<Mutability>

Get the first argument’s mutability (if it is a pointer)

pub fn simplify_until_type_checks(&self, env: &Env) -> Result<Self, Error>

Simplify until the type passes the type checker.

pub fn simplify_until_has_members(&self, env: &Env) -> Result<Self, Error>

Simplify a type until you can get its members.

pub fn simplify_until_union(&self, env: &Env) -> Result<Self, Error>

Simplify a type until it’s a union.

pub fn simplify_until_has_variants(&self, env: &Env) -> Result<Self, Error>

Simplify a type until you can get its variants.

pub fn simplify_until_atomic(&self, env: &Env) -> Result<Self, Error>

pub fn simplify_until_simple(&self, env: &Env) -> Result<Self, Error>

pub fn simplify_until_poly(&self, env: &Env) -> Result<Self, Error>

Simplify until the type is a polymorphic type.

pub fn simplify_until_concrete(&self, env: &Env) -> Result<Self, Error>

Simplify until the type is concrete.

pub fn simplify_until_matches( self, env: &Env, expected: Self, f: impl Fn(&Self, &Env) -> Result<bool, Error> ) -> Result<Self, Error>

Simplify an expression until it matches a given function which “approves” of a type. This will perform template applications to simplify the type if possible as well.

This is usually too verbose to be used on its own, better to make a wrapper function that calls this with a more specific use case/purpose.

pub fn let_bind(name: &str, t: Self, ret: Self) -> Self

Create a let-bound type.

pub fn variant_index(variants: &[String], variant: &String) -> Option<usize>

Calculate the integral value of a variant in an enum.

pub fn contains_symbol(&self, name: &str) -> bool

Does this type contain a symbol with the given name? This will not count overshadowded versions of the symbol (overwritten by let-bindings).

pub fn substitute(&self, name: &str, substitution: &Self) -> Self

Substitute all occurences of a symbol with another type. This will not traverse into let-bindings when the symbol is overshadowed.

pub fn has_element_type(&self, element: &Self, env: &Env) -> Result<bool, Error>

Does this type have an element type matching the supplied type? If this type is an array of Integers, for example, then this function will return true if the supplied type is an Integer.

pub fn can_decay_to(&self, desired: &Self, env: &Env) -> Result<bool, Error>

Can this type decay into another type?

A type can decay into another type if the compiler can automatically convert the type into the other type. This is distinct from the compiler seeing the types as equal. The two types are not equal; the compiler will just automatically perform the conversion.

For all cases right now, a decay does nothing; the representations of the types in the compiler are the same for all types of decay.

pub fn can_cast_to(&self, other: &Self, env: &Env) -> Result<bool, Error>

Can this type be cast to another type?

pub fn equals(&self, other: &Self, env: &Env) -> Result<bool, Error>

Are two types structurally equal?

pub fn perform_template_applications( &self, env: &Env, previous_applications: &mut HashMap<(Type, Vec<Type>), Type> ) -> Result<Self, Error>

Perform type applications if possible.