Enum sage::lir::Type

pub enum Type {

    Let(String, Box<Self>, Box<Self>),
    Unit(String, Box<Self>),
    Symbol(String),
    None,
    Int,
    Float,
    Cell,
    Char,
    Bool,
    Enum(Vec<String>),
    Tuple(Vec<Self>),
    Array(Box<Self>, Box<ConstExpr>),
    Struct(BTreeMap<String, Self>),
    Union(BTreeMap<String, Self>),
    Proc(Vec<Type>, Box<Type>),
    Pointer(Box<Self>),
    Any,
    Never,

}

The representation of a type in the LIR type system.

Variants

Let(String, Box<Self>, Box<Self>)

Bind a type to a name in a temporary scope.

Unit(String, Box<Self>)

This type is identified by its name. Most types are checked according to structural equality, but this type is checked according to name. Structural equality is also verified in addition to name equality.

The inner type acts exactly the same, but it can only be type-equal to another Unit type with the same name and inner type.

Symbol(String)

A named type.

None

The type of void expressions.

Int

The integer type.

Float

The floating-point number type.

Cell

The type of the most basic unit of memory.

Char

The type of a character.

Bool

The type of a boolean value.

Enum(Vec<String>)

An enumeration of a list of possible named values. A boolean could be considered an enumeration of true and false.

Tuple(Vec<Self>)

A heterogenous collection of types. This is a product type.

Array(Box<Self>, Box<ConstExpr>)

An array of a given type, with a constant size.

Struct(BTreeMap<String, Self>)

A tuple with named members. This is a product type.

Union(BTreeMap<String, Self>)

A union of a list of possible types mapped to named members. A union’s value is reinterpreted as a single type, depending on the member accessed. Unions’ values are stored starting at the beginning of the union’s address in memory, and are padded at the end with zeroes.

This is a sum type.

Proc(Vec<Type>, Box<Type>)

A procedure with a list of parameters and a return type.

Pointer(Box<Self>)

A pointer to another type.

Any

A type reserved by the compiler. This type is equal to any other type. The NULL pointer, for example, is of type Pointer(Any).

Never

The type of an expression that will never return, or doesn’t resolve to a value.

Implementations

impl Type

pub fn variant_index(variants: &Vec<String>, variant: &String) -> Option<usize>

Calculate the integral value of a variant in an enum.

pub fn contains_symbol(&self, name: &str) -> bool

Does this type contain a symbol with the given name? This will not count overshadowded versions of the symbol (overwritten by let-bindings).

pub fn substitute(&self, name: &str, substitution: &Self) -> Self

Substitute all occurences of a symbol with another type. This will not traverse into let-bindings where the symbol is overshadowed.

pub fn can_cast_to(&self, other: &Self, env: &Env) -> Result<bool, Error>

Can this type be cast to another type?

pub fn equals(&self, other: &Self, env: &Env) -> Result<bool, Error>

Are two types structurally equal?