Struct sage::lir::Env

pub struct Env { /* private fields */ }

An environment under which expressions and types are compiled and typechecked. This is essentially the scope of an expression.

Implementations

impl Env

pub fn define_types(&mut self, types: Vec<(String, Type)>)

Define multiple types with the given names under this environment.

This must be used in situations where the different types depend on each other. This is because the sizes of types are memoized, and this will interfere with the memoization process if the types are defined separately. It will lead to typechecking errors if the environment does not already have a memoized size for the type of a subexpression.

pub fn define_var( &mut self, var: impl ToString, mutability: Mutability, ty: Type ) -> Result<isize, Error>

Define a variable in the current scope. This will increment the scope’s frame pointer offset by the size of the variable. This method returns the offset of the variable from the frame pointer under this scope.