Trait sage::lir::AssignOp

pub trait AssignOp: Debug + Display {
    fn clone_box(&self) -> Box<dyn AssignOp>;
    fn can_apply(&self, dst: &Type, src: &Type, env: &Env) -> Result<bool, Error>;
    fn eval(
        &self,
        dst: &ConstExpr,
        src: &ConstExpr,
        env: &mut Env
    ) -> Result<ConstExpr, Error>;
    fn compile_types(
        &self,
        dst: &Type,
        src: &Type,
        env: &mut Env,
        output: &mut dyn AssemblyProgram
    ) -> Result<(), Error>;

    fn type_check(&self, dst: &Expr, src: &Expr, env: &Env) -> Result<(), Error> { ... }
    fn return_type(&self, dst: &Expr, src: &Expr, env: &Env) -> Result<Type, Error> { ... }
    fn display(&self, dst: &Expr, src: &Expr) -> String { ... }
    fn can_apply_exprs(
        &self,
        dst: &Expr,
        src: &Expr,
        env: &Env
    ) -> Result<bool, Error> { ... }
    fn compile(
        &self,
        dst: &Expr,
        src: &Expr,
        env: &mut Env,
        output: &mut dyn AssemblyProgram
    ) -> Result<(), Error> { ... }
}

A trait used to implemented an assignment operation.

This trait is used to implement assignment operations like += and -=.

Required Methods

fn clone_box(&self) -> Box<dyn AssignOp>

Clones the operation into a boxed trait object.

fn can_apply(&self, dst: &Type, src: &Type, env: &Env) -> Result<bool, Error>

Checks if the operation can be applied to the given types.

fn eval(
&self,
dst: &ConstExpr,
src: &ConstExpr,
env: &mut Env
) -> Result<ConstExpr, Error>

Evaluates the operation on the given constant expressions.

fn compile_types(
&self,
dst: &Type,
src: &Type,
env: &mut Env,
output: &mut dyn AssemblyProgram
) -> Result<(), Error>

Compiles the operation on the given types. (Generates the code for the operation.)

Provided Methods

fn type_check(&self, dst: &Expr, src: &Expr, env: &Env) -> Result<(), Error>

Typechecks the operation on the given expressions.

fn return_type(&self, dst: &Expr, src: &Expr, env: &Env) -> Result<Type, Error>

Gets the type of the operation on the given expressions.

fn display(&self, dst: &Expr, src: &Expr) -> String

Formats the operation for display.

fn can_apply_exprs(&self, dst: &Expr, src: &Expr, env: &Env) -> Result<bool, Error>

Checks if the operation can be applied to the given expressions.

fn compile(
&self,
dst: &Expr,
src: &Expr,
env: &mut Env,
output: &mut dyn AssemblyProgram
) -> Result<(), Error>

Compiles the operation on the given expressions.