Trait sage::lir::Compile

pub trait Compile: TypeCheck + Debug + Display {
    // Required method
    fn compile_expr(
        self,
        env: &mut Env,
        output: &mut dyn AssemblyProgram
    ) -> Result<(), Error>;

    // Provided method
    fn compile(self) -> Result<Result<CoreProgram, StandardProgram>, Error>
       where Self: Sized + Clone { ... }
}

A trait which allows an LIR expression to be compiled to one of the two variants of the assembly language.

Required Methods

fn compile_expr( self, env: &mut Env, output: &mut dyn AssemblyProgram ) -> Result<(), Error>

Provided Methods

fn compile(self) -> Result<Result<CoreProgram, StandardProgram>, Error>
where Self: Sized + Clone,

Compile the expression into an assembly program.

On success, this will return an Ok value containing either a successfully compiled core assembly program, or a fallback standard assembly program.

On an error, this will return an Err value containing the error.