Trait sage::lir::GetSize

pub trait GetSize {
    fn get_size_checked(&self, env: &Env, i: usize) -> Result<usize, Error>;

    fn get_size(&self, env: &Env) -> Result<usize, Error> { ... }
}

Get the size of something in memory (number of cells).

Required Methods

fn get_size_checked(&self, env: &Env, i: usize) -> Result<usize, Error>

Get the size of something in memory, but limit the number of recursive calls to prevent stack overflow. i is a counter to prevent infinite recursion.

Provided Methods

fn get_size(&self, env: &Env) -> Result<usize, Error>

Get the size of something in memory (number of cells).