Enum sage::side_effects::io::InputMode

pub enum InputMode {

    StdinChar,
    StdinInt,
    StdinFloat,
    Random,
    DPad(Direction),
    Button,
    Keyboard,
    JoyStick(Direction),
    Clock,
    Accelerometer(Axis),
    Gyroscope(Axis),
    Microphone,
    RedLight,
    GreenLight,
    BlueLight,
    Brightness,
    Humidity,
    Barometer,
    Magnetometer(Axis),
    Thermometer,
    RainGauge,
    UVSensor,
    WindSpeed,
    WindDirection,
    PressureGauge,
    FlowSensor,
    VolumeSensor,
    WeightSensor,
    PHSensor,
    ConductivitySensor,
    Speedometer(Option<Axis>),
    Odometer,
    Position(Axis),
    Compass,
    Proximity,
    Altimeter,
    DepthSensor,
    AnalogPin,
    DigitalPin,
    Custom(String),

}

The different types of input modes a program might use.

Variants

StdinChar

Standard input modes (The standard interface is typically used for command line programs, and for all pipelines between programs) Standard input (ASCII character)

StdinInt

Standard input (integer)

StdinFloat

Standard input (float)

Random

Special input modes A random number

DPad(Direction)

User input modes (These should typically be used for games or other interactive programs) Input from a D-Pad

Button

Input from a button (0=not pressed, 1=pressed)

Keyboard

Input from keyboard (ASCII character)

JoyStick(Direction)

Input from a JoyStick the degree of displacement in a given direction (from -128 to 128).

Clock

Physical sensor input modes (These should typically be implemented for devices that measure the physical world, like a smart watch or a smart phone) Time (in seconds) since the program started

Accelerometer(Axis)

Input from an accelerometer (in meters per second per second) in a given axis

Gyroscope(Axis)

Input from a gyroscope (in degrees per second) around a given axis

Microphone

Input from a microphone (frequency in hertz)

RedLight

Environment sensor input modes (These should typically be implemented for devices that measure the environment, like a weather station or a smart thermostat) Red light intensity (in lux)

GreenLight

Green light intensity (in lux)

BlueLight

Blue light intensity (in lux)

Brightness

Input from a light sensor (in lux)

Humidity

Input from a humidity sensor (in percent)

Barometer

Input from a barometer (pressure in atmospheres)

Magnetometer(Axis)

Input from a magnetometer (in teslas) in a given axis

Thermometer

Input from a thermometer (degrees K)

RainGauge

Input from a rain gauge (in millimeters)

UVSensor

Input from a UV sensor (in watts per square meter)

WindSpeed

Input from a wind speed sensor (in meters per second)

WindDirection

Input from a wind direction sensor (in degrees)

PressureGauge

Engineering / Science sensor input modes Input from a pressure gauge (in atmospheres)

FlowSensor

Input from a flow sensor (in liters per second)

VolumeSensor

Input from a volume sensor (in liters)

WeightSensor

Input from a weight sensor (in kilograms)

PHSensor

Input from a pH sensor (in pH)

ConductivitySensor

Input from a conductivity sensor (in siemens per meter)

Speedometer(Option<Axis>)

Navigation input modes (These should typically be implemented for devices that use GPS, or for robots) Input from a speedometer (in meters per second) in a given axis

Odometer

Input from an odometer (in meters)

Position(Axis)

Input from a position sensor in a given axis (x, y, z)

Compass

Input from a compass (degrees)

Proximity

Input from a distance sensor (in meters)

Altimeter

Input from altitude sensor (in meters)

DepthSensor

Input from depth sensor (in meters)

AnalogPin

Electrical device input modes (These should typically be used for programs using a GPIO interface, like a Raspberry Pi or Arduino) Input from an analog input (in volts)

DigitalPin

Input from a digital input (0=low, 1=high)

Custom(String)

Custom output modes A custom input mode (for use with a custom output device)