Trait sage::vm::Device

pub trait Device {
    fn get(&mut self, src: Input) -> Result<i64, String>;
    fn put(&mut self, val: i64, dst: Output) -> Result<(), String>;
    fn peek(&mut self) -> Result<i64, String>;
    fn poke(&mut self, val: i64) -> Result<(), String>;
}

Create an input / output device for the virtual machine interpreter to operate on. The method get retrieves the device’s input, and the function put writes to the devices output.

TODO: Make a trait for a device with the standard variant, which requires get_char, put_char, get_int, put_int, get_float, and put_float methods.

Required Methods

fn get(&mut self, src: Input) -> Result<i64, String>

Get the next input (from a given input source).

fn put(&mut self, val: i64, dst: Output) -> Result<(), String>

Put the given value to the given output destination.

fn peek(&mut self) -> Result<i64, String>

Peek a value from the side-effecting device wrapping the virtual machine.

fn poke(&mut self, val: i64) -> Result<(), String>

Poke a value into the side-effecting device wrapping the virtual machine.